Refs #107. slight refactor to move autocrop funcs to its own module
