+++ python/imars3d/CTProcessor.py
@@ -341,46 +341,10 @@ The default behavior can be modified by configuration file "imars3d.conf".
 
     @dec.timeit
     def autoCrop(self, series):
-        # estimate average
-        ave = self.estimateAverage(series)
-        from . import io
-        def save(d, p): 
-            im = io.ImageFile(p); im.data = d; im.save()
-        save(ave, "estimate-ave.tiff")
-        # smooth it
-        from scipy import ndimage 
-        sm = ndimage.median_filter(ave, 9)
-        save(sm, "sm-estimate-ave.tiff")
-        # find foreground rectangle
-        Y, X = np.where(sm < 0.8)
-        ymax = Y.max(); ymin = Y.min()
-        xmax = X.max(); xmin = X.min()
-        # expand it a bit
-        width = xmax - xmin; height = ymax - ymin
-        xmin -= width/6.; xmax += width/6.
-        ymin -= height/6.; ymax += height/6.
-        HEIGHT, WIDTH = ave.shape
-        if xmin < 0: xmin = 0
-        if xmax > WIDTH-1: xmax =WIDTH-1
-        if ymin < 0: ymin = 0
-        if ymax > HEIGHT-1: ymax =HEIGHT-1
-        xmin, xmax, ymin, ymax = map(int, (xmin, xmax, ymin, ymax))
+        from .autocrop import calculateCropWindow
+        xmin, xmax, ymin, ymax = calculateCropWindow(series)
         # crop
         return self.crop(series, left=xmin, right=xmax, top=ymin, bottom=ymax)
-
-
-    def estimateAverage(self, series):
-        sum = None; N = 0
-        for i,img in enumerate(series):
-            if i%5!=0: continue # skip some
-            data = img.data
-            if sum is None:
-                sum = np.array(data, dtype='float32')
-            else:
-                sum += data
-            N += 1
-            continue
-        return sum/N
 
 
     def crop(self, series, left=None, right=None, top=None, bottom=None):
